[PATCH] cdev 1/2: Eliminate /sys/cdev
This is the first of two patches designed to make life easier for authors
of device driver books - and, with luck, driver authors too.

/sys/cdev is, according to Al, a mistake which was never really meant to
exist.  I believe nothing uses it currently - there isn't a whole lot there
to use.  Its existence takes up system resources, and also requires drivers
to deal with the cdev's embedded kobject in their failure paths.  The
following patch (against 2.6.4) makes it all go away.

OK, almost all.  We have to keep (but not register) cdev_subsys because
it's rwsem is needed to control access to cdev_map.