[PATCH] add device quiescing to the SCSI API
This patch adds the ability to quiesce a SCSI device.  The idea is that
user issued commands (including filesystem ones) would get blocked,
while mid-layer and device issued ones would be allowed to proceed.
This is for things like Domain Validation which like to operate on an
otherwise quiet device.

There is one big change: to get all of this to happen correctly,
scsi_do_req() has to queue on the *head* of the request queue, not the
tail as it was doing previously.  The reason is that deferred requests
block the queue, so anything needing executing after a deferred request
has to go in front of it.  I don't think there are any untoward
consequences of this.