[SPARC32]: Rework the CPU enumeration and probing code
The low 2 bits of the MID are now used exclusively as the software
CPU ID on 4m.  This eliminates entirely the need for mid_xlate.  We
also eliminate linux_num_cpus and linux_cpus; since the hardware-based
id is now used everywhere we don't need indexed CPU tables.

The device probe is almost identical to sparc64's version.  The one
exception is that 4m only uses the bottom 2 bits of the MID, so we also
have a separate function to get the true hardware MID for the rare times
it's actually needed.

Per-cpu data is now used the same way as in sparc64.  As a side benefit
we now display the same format Bogo and Clock values in /proc/cpuinfo.