ia64: GCC v3.5 fixes.
GCC v3.5 doesn't like it when a function gets cast to a
function-pointer with a different signature.  In fact, it exterts
extreme prejudice in this case and replaces the call-site with a call
to __builtin_trap()!  Fix this by providing appropriately typed
dummy-functions for the macvec operations as needed.

While at it, also clean up some machvec handlers which weren't used anywhere.