[PATCH] ip2: fix double operator
From: "Randy.Dunlap" <rddunlap@osdl.org>

This trivial patch fixes the bug #320:

http://bugme.osdl.org/show_bug.cgi?id=320

The additional comment:

http://bugme.osdl.org/show_bug.cgi?id=320#c1

Is wrong, because it will send the command twice.

We only want to see if the command got success, thus is not necessary
to test against < 0 (if the return value is not 1, we got a error).

Note that I'm using the function (bad) style.