[PATCH] add note about "Copyright" to SubmittingDrivers
From: Grant Grundler <grundler@parisc-linux.org>

This patch adds a comment to "Documentation/SubmittingDrivers" about the
importance of adding a Copyright notice in submitted code.

The parisc-linux port has neglected this in the past and I've been slowly
trying to correct that (along with proper GPL header).

While I make it sound like GPL is the "only" acceptable license, I'll leave
it up to lawyers to determine what other appropriate license could be used
for a new driver.