[CRYPTO]: Add setkey operation for digests.
From Jouni Malinen <jkmaline@cc.hut.fi>

Added support for using keyed digest with an optional dit_setkey handler.
This does not change the behavior of the existing digest algorithms, but
allows new ones to add setkey handler that can be used to initialize the
algorithm with a key or seed. setkey is to be called after init, but before
any of the update call(s).