[PATCH] fbcon font cloning fix
From: Petr Baudis <pasky@ucw.cz>

This patch fixes a bug which I hit when migrating from 2.4.  Basically,
when I do setfont during the system boot, the then-spawned ttys do not
retain the new font but fall back to the default one.

I've tracked that down to a clearly bogus test in fbcon_set_display(),
because vc->vc_font.width is not set at that time yet (no font has been
loaded for the new vc).  But even if it would (or this was meant to test
against tmp->vc_font.width), it would mean only *HUGE* fonts would be
retained.  And even if there were *two* bugs there and it was supposed to
be less-than there, I couldn't make a sense of the test.