[PATCH] fix ppc32 sys_swapcontext
This fixes a bug in the swapcontext system call on ppc32.

On ppc32, the system call entry only saves the volatile registers,
except in the case of a few system calls (e.g.  fork) which need all the
registers saved.  Swapcontext needs all the registers but we weren't
saving them all. So fixes that.