[PATCH] USB: Eliminate wait following interface unregistration
This patch from Alan Stern <stern@rowland.harvard.edu> fixes a bug in
the current USB code that causes khubd to hang when a device is removed
from the system, thereby preventing any future USB device changes (like
adding or removing other devices) from happening.

Both Andrew and I can easily duplicate this bug against the current -bk
tree. 

It's not a perfect fix, but it works for now, and I will spend the next
week working on restructuring the code so this is handled properly.