[PATCH] USB Gadget: Use configuration-buffer library in file-storage
This patch imports the config-buffer library into the file-storage gadget,
simplifying and decreasing the amount of code needed for assembling
configuration descriptors.  It also changes the driver to remove any
pretense at bus-powered operation and to use the new DUALSPEED
configuration option.  This is in line with recent changes made to other
gadget drivers.